font added
Former-commit-id: dd8bacc666a798ee6343f8d535b7b545e46b6f41
Former-commit-id: d71b1a86212cb2093880c3ffdf87e5c65e52dc9c
Former-commit-id: 916013df91ecfdbc087143e4311b6e055fa8a82a
Former-commit-id: 30ec9c8ea1086caa71aac7a1bb23170c30c5231e

+++ scripts/installCfg.sh
@@ -89,7 +89,10 @@ mkdir -p ~/pictures
 ${RS[@]} "${RES}/wallpapers" ~/pictures >> $LOG
 
 ## Adapta GTK Colorthemes
-sudo dpkg --install "${RES}/others/AdaptaGTKcolorpack3-94-0-149.deb"
+pkgs='adapta-gtk-theme-colorpack'
+if ! dpkg -s $pkgs >/dev/null 2>&1; then
+    sudo dpkg --install "${RES}/others/AdaptaGTKcolorpack3-94-0-149.deb"
+fi
 
 # ? Reload of services and caches
